requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
passlib[bcrypt]
python-multipart
PyPDF2
sentence-transformers
faiss-cpu
numpy
